// src/models/transaction.entity.ts
export class Transaction {
  id: number;

  transaction_id: string;

  
  booking_date: string;

  
  value_date: string;

  
  remittance_info: string;

  
  reference: string;

  
  discount: number;

  
  amount: number;

  
  posted: boolean;

  
  credit_debit_indicator: string;

  // Foreign key definition
  
  batch_id: number;


  
}